Move src/types from a src/lib/structs
