+++ webapp/src/components/connectorform/FormSection.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import { Schema } from 'types/connectorform/form';
+import { Schema } from 'struct/form';
 
 import FormField from './FormField';
