+++ webapp/src/lib/struct/agent.ts
@@ -7,23 +7,69 @@ export type CodeExecutionConfigType = {
 	lastNMessages: number;
 	workDirectory: string;
 };
-
 export type Agent = {
-	_id?: ObjectId;
-	orgId?: ObjectId;
-	teamId?: ObjectId;
+	_id?: ObjectId | string;
+	orgId?: ObjectId | string;
+	teamId?: ObjectId | string;
 	name: string;
 	role: string;
 	goal: string;
 	backstory: string;
-	modelId: ObjectId;
-	functionModelId: ObjectId;
-	maxIter: number;
-	maxRPM: number;
+	modelId: ObjectId | string;
+	functionModelId: ObjectId | string | null;
+	maxIter: number | null;
+	maxRPM: number | null;
 	verbose: boolean;
 	allowDelegation: boolean;
-	toolIds?: ObjectId[];
-	icon?: IconAttachment;
-	// stepCallback: Function;
+	toolIds?: (ObjectId | string)[];
+	icon?:
+		| IconAttachment
+		| {
+				id: string;
+				filename: string;
+		  };
 	hidden?: boolean;
+	group?: any[];
 };
+
+// export type Agent = {
+// 	_id?: ObjectId;
+// 	orgId?: ObjectId;
+// 	teamId?: ObjectId;
+// 	name: string;
+// 	role: string;
+// 	goal: string;
+// 	backstory: string;
+// 	modelId: ObjectId;
+// 	functionModelId: ObjectId;
+// 	maxIter: number;
+// 	maxRPM: number;
+// 	verbose: boolean;
+// 	allowDelegation: boolean;
+// 	toolIds?: ObjectId[];
+// 	icon?: IconAttachment;
+// 	// stepCallback: Function;
+// 	hidden?: boolean;
+// };
+
+// export interface Agent1 {
+// 	_id: string;
+// 	orgId: string;
+// 	teamId: string;
+// 	name: string;
+// 	role: string;
+// 	goal: string;
+// 	backstory: string;
+// 	modelId: string;
+// 	functionModelId: string | null;
+// 	maxIter: number | null;
+// 	maxRPM: number | null;
+// 	verbose: boolean;
+// 	allowDelegation: boolean;
+// 	toolIds: string[];
+// 	icon: {
+// 		id: string;
+// 		filename: string;
+// 	};
+// 	group: any[];
+// }
