--- webapp/src/types/app/index.ts
@@ -1,2 +1,0 @@
-export * from './agent';
-export * from './task';
